requirements.txt
requests
python-dotenv
scikit-learn
matplotlib
pandas
plotly
seaborn